docs: table complex filters fixes
