+++ .storybook/preview.tsx
@@ -1,9 +1,16 @@
 import type { Preview } from '@storybook/react'
+// import type { IndexEntry } from '@storybook/types'
 
 const preview: Preview = {
   parameters: {
     layout: 'centered',
     options: {
+      // TODO https://mparticle-eng.atlassian.net/browse/UNI-1214
+      // storySort: (a: IndexEntry, b: IndexEntry) => {
+      // console.log('Ordering stories', { a, b })
+      // const order = ['Documentation', 'Cell Types', 'Filters', 'Primary', 'Complex']
+      // return order.indexOf(a[1].name) - order.indexOf(b[1].name)
+      // },
       storySort: {
         order: [
           'About',
@@ -38,7 +45,7 @@ const preview: Preview = {
           ],
           'Candidate Components',
           'UX Patterns',
-          ['Table', ['Table', ['Documentation', 'Cell Types', 'Basic Table'], 'Filters']],
+          ['Table', ['Table', ['Documentation', 'Cell Types', 'Filters', 'Primary', 'Complex']]],
           'Contributing',
           ['Introduction', 'Commits', 'Testing in the platforms', 'Release Process', 'Maintainers'],
         ],
