+++ src/components/data-display/Collapse/Collapse.tsx
@@ -1,5 +1,5 @@
 import { Collapse as AntCollapse } from 'antd'
-import { type CollapseProps as AntCollapseProps } from 'antd'
+import type { CollapseProps as AntCollapseProps } from 'antd'
 import { ConfigProvider } from 'src/components'
 
 export interface ICollapseProps extends AntCollapseProps {}
